Fixes to benchmarks and greater backward compatibility
