+++ zio-json/jvm/src/jmh/scala/zio/json/SyntheticBenchmarks.scala
@@ -2,7 +2,6 @@ package zio.json
 
 import java.nio.charset.StandardCharsets.UTF_8
 import java.util.concurrent.TimeUnit
-import zio.json
 
 import com.github.plokhotnyuk.jsoniter_scala.core._
 import com.github.plokhotnyuk.jsoniter_scala.macros._
