+++ zio-json/jvm/src/jmh/scala/zio/json/TwitterAPIBenchmarks.scala
@@ -3,7 +3,6 @@ package zio.json
 import java.nio.charset.StandardCharsets.UTF_8
 import java.util.concurrent.TimeUnit
 
-import zio.json
 import com.github.plokhotnyuk.jsoniter_scala.core._
 import com.github.plokhotnyuk.jsoniter_scala.macros._
 import io.circe
